CLEAN: Moved material items for projects, certificates, and blogs into one directory - Easier to find
